# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /psychology/analysis:
    post:
      operationId: analyzePsychology
      tags: [psychology]
      summary: Build a psychological profile from the user's goals
      description: >
        Takes the user's goals (with per-goal completion stats) and returns a
        psychological profile: goals grouped by their nature into categories,
        each with an accurate completion rate, plus narrative traits and insights.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PsychAnalysisInput"
      responses:
        "200":
          description: Analysis produced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PsychAnalysis"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
  /psychology/plan:
    post:
      operationId: planPsychology
      tags: [psychology]
      summary: Turn the goal profile into a concrete action plan
      description: >
        Takes the user's profile (summary, traits, categories, insights) plus their
        goals, reflections and context, and — accepting the read at face value —
        returns a small set of concrete, prioritised moves the user should make.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PsychPlanInput"
      responses:
        "200":
          description: Action plan produced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PsychPlan"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
  /psychology/chat:
    post:
      operationId: chatPsychology
      tags: [psychology]
      summary: Discuss the goal profile with the assistant
      description: >
        Answers the user's questions grounded in their goals, completion-by-category
        data and profile. Supports hypotheticals such as proposing a new goal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PsychChatInput"
      responses:
        "200":
          description: Assistant reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PsychChatReply"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
...
components:
  schemas:
...
    ApiErrorMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    GoalSnapshot:
      type: object
      description: A single goal plus its observed completion stats.
      properties:
        title:
          type: string
        notes:
          type: ["string", "null"]
        timeframe:
          type: string
          description: One of daily, medium, long.
        importance:
          type: ["integer", "null"]
        done:
          type: number
          description: Credited completions (partial counts as 0.5).
        due:
          type: integer
          description: Number of tracked occurrences that were due.
        rate:
          type: number
          description: done / due, between 0 and 1.
      required:
        - title
        - timeframe
        - done
        - due
        - rate
    PsychTrait:
      type: object
      properties:
        label:
          type: string
        score:
          type: integer
          description: 0-100 strength of this trait.
        note:
          type: string
      required:
        - label
        - score
        - note
    PsychCategory:
      type: object
      properties:
        name:
          type: string
        blurb:
          type: string
        taskCount:
          type: integer
        done:
          type: number
        due:
          type: integer
        rate:
          type: number
      required:
        - name
        - blurb
        - taskCount
        - done
        - due
        - rate
    ReflectionEntry:
      type: object
      properties:
        period:
          type: string
          description: One of day, week, month, year.
        label:
          type: string
          description: Human-readable period label, e.g. "Monday, June 9, 2026" or "May 2026".
        text:
          type: string
          description: What the user reported actually accomplishing in that period.
      required:
        - period
        - label
        - text
    PsychAnalysisInput:
      type: object
      properties:
        goals:
          type: array
          items:
            $ref: "#/components/schemas/GoalSnapshot"
        reflections:
          type: array
          description: The user's own free-text accounts of what they accomplished, by period.
          items:
            $ref: "#/components/schemas/ReflectionEntry"
        context:
          type: ["string", "null"]
          description: The user's own free-text context — their side of the story, things the stats and reflections don't capture, which the model may weigh when profiling.
      required:
        - goals
    PsychAnalysis:
      type: object
      properties:
        generatedAt:
          type: string
        headline:
          type: string
        summary:
          type: string
        traits:
          type: array
          items:
            $ref: "#/components/schemas/PsychTrait"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/PsychCategory"
        insights:
          type: array
          items:
            type: string
      required:
        - generatedAt
        - headline
        - summary
        - traits
        - categories
        - insights
    PsychChatMessage:
      type: object
      properties:
        role:
          type: string
          description: user or assistant.
        content:
          type: string
      required:
        - role
        - content
    PsychChatInput:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/PsychChatMessage"
        goals:
          type: array
          items:
            $ref: "#/components/schemas/GoalSnapshot"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/PsychCategory"
        reflections:
          type: array
          description: The user's own free-text accounts of what they accomplished, by period.
          items:
            $ref: "#/components/schemas/ReflectionEntry"
        context:
          type: ["string", "null"]
          description: The user's own free-text context — their side of the story, things the stats and reflections don't capture, which the model may weigh when replying.
        profileSummary:
          type: ["string", "null"]
      required:
        - messages
        - goals
    PsychChatReply:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply
    PsychPlanInput:
      type: object
      properties:
        goals:
          type: array
          items:
            $ref: "#/components/schemas/GoalSnapshot"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/PsychCategory"
        traits:
          type: array
          items:
            $ref: "#/components/schemas/PsychTrait"
        insights:
          type: array
          items:
            type: string
        headline:
          type: ["string", "null"]
        profileSummary:
          type: ["string", "null"]
        reflections:
          type: array
          description: The user's own free-text accounts of what they accomplished, by period.
          items:
            $ref: "#/components/schemas/ReflectionEntry"
        context:
          type: ["string", "null"]
          description: The user's own free-text context — their side of the story, which the model may weigh.
      required:
        - goals
    PsychPlanMove:
      type: object
      properties:
        title:
          type: string
          description: A short imperative name for the move.
        detail:
          type: string
          description: What to do concretely and why it follows from the profile.
      required:
        - title
        - detail
    PsychPlan:
      type: object
      properties:
        generatedAt:
          type: string
        premise:
          type: string
          description: One or two sentences restating the read being taken at face value.
        moves:
          type: array
          items:
            $ref: "#/components/schemas/PsychPlanMove"
        firstStep:
          type: string
          description: The single thing to do first, today.
      required:
        - generatedAt
        - premise
        - moves
        - firstStep
...