# lib/api-zod/src/generated/types/psychAnalysisInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GoalSnapshot } from './goalSnapshot';
import type { ReflectionEntry } from './reflectionEntry';

export interface PsychAnalysisInput {
  goals: GoalSnapshot[];
  /** The user's own free-text accounts of what they accomplished, by period. */
  reflections?: ReflectionEntry[];
  /**
     * The user's own free-text context — their side of the story, things the stats and reflections don't capture, which the model may weigh when profiling.
     * @nullable
     */
  context?: string | null;
}