# lib/api-zod/src/generated/types/psychChatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GoalSnapshot } from './goalSnapshot';
import type { PsychCategory } from './psychCategory';
import type { PsychChatMessage } from './psychChatMessage';
import type { ReflectionEntry } from './reflectionEntry';

export interface PsychChatInput {
  messages: PsychChatMessage[];
  goals: GoalSnapshot[];
  categories?: PsychCategory[];
  /** The user's own free-text accounts of what they accomplished, by period. */
  reflections?: ReflectionEntry[];
  /**
     * The user's own free-text context — their side of the story, things the stats and reflections don't capture, which the model may weigh when replying.
     * @nullable
     */
  context?: string | null;
  /** @nullable */
  profileSummary?: string | null;
}