# lib/api-zod/src/generated/types/psychPlan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PsychPlanMove } from './psychPlanMove';

export interface PsychPlan {
  generatedAt: string;
  /** One or two sentences restating the read being taken at face value. */
  premise: string;
  moves: PsychPlanMove[];
  /** The single thing to do first, today. */
  firstStep: string;
}