# lib/api-zod/src/generated/types/psychPlanInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GoalSnapshot } from './goalSnapshot';
import type { PsychCategory } from './psychCategory';
import type { PsychTrait } from './psychTrait';
import type { ReflectionEntry } from './reflectionEntry';

export interface PsychPlanInput {
  goals: GoalSnapshot[];
  categories?: PsychCategory[];
  traits?: PsychTrait[];
  insights?: string[];
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  profileSummary?: string | null;
  /** The user's own free-text accounts of what they accomplished, by period. */
  reflections?: ReflectionEntry[];
  /**
     * The user's own free-text context — their side of the story, which the model may weigh.
     * @nullable
     */
  context?: string | null;
}