# lib/api-zod/src/generated/types/psychPlanMove.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PsychPlanMove {
  /** A short imperative name for the move. */
  title: string;
  /** What to do concretely and why it follows from the profile. */
  detail: string;
}