# lib/api-zod/src/generated/types/reflectionEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReflectionEntry {
  /** One of day, week, month, year. */
  period: string;
  /** Human-readable period label, e.g. "Monday, June 9, 2026" or "May 2026". */
  label: string;
  /** What the user reported actually accomplishing in that period. */
  text: string;
}